utils support dynamic modelName
